Watch old CRD group to migrate CR to new group
After removing all of its finalizers, we delete the old CR and create
a new one from its Config. The k8s GC should ensure all its owned
resources are deleted. The new CR should immediately re-create them.
We may need to revisit this if that service interruption isn't
acceptable, but if so, we'll need to account for the previous
controller's ownership of those existing resources.